+++ src/App.tsx
@@ -3,31 +3,31 @@ import HomePage from "./pages/HomePage";
 import RecsPage from "./pages/RecsPage";
 import SearchPage from "./pages/SearchPage";
 import FavoritesPage from "./pages/FavoritesPage";
-import MyPage from "./pages/MyPage";
+import MyPage from "./pages/myPage/MyPage";
 import SpaceDetailPage from "./pages/SpaceDetailPage";
 import ThemeAllPage from "./pages/ThemeAllPage";
 import ThemeDetailPage from "./pages/ThemeDetailPage";
 import SignupPage from "./pages/SignupPage";
-import NoticePage from "./pages/NoticePage";
-import NoticeDetailPage from "./pages/NoticeDetailPage";
+import NoticePage from "./pages/myPage/NoticePage";
+import NoticeDetailPage from "./pages/myPage/NoticeDetailPage";
 import PointLayout from "./layouts/PointLayout";
 import MainList from "./components/myPage/point/MainList";
 import PointCharge from "./components/myPage/point/PointCharge";
 import PointDetail from "./components/myPage/point/PointDetail";
 import HotSpaceListPage from "./pages/HotSpaceListPage";  
-import ProfilePage from "./pages/ProfilePage";
-import WithdrawalPage from "./pages/WithdrawalPage";
-import ProposalPage from "./pages/ProposalPage";
+import ProfilePage from "./pages/myPage/ProfilePage";
+import WithdrawalPage from "./pages/myPage/WithdrawalPage";
+import ProposalPage from "./pages/myPage/ProposalPage";
 import SpaceReviewWritePage from "./pages/SpaceReviewWritePage";
 import LoginPage from "./pages/LoginPage";
 import OauthKakaoCallback from "./components/login/OauthKakaoCallback";
-import PushPage from "./pages/PushPage";
-import VisitPage from "./pages/VisitPage";
+import PushPage from "./pages/myPage/PushPage";
+import VisitPage from "./pages/myPage/VisitPage";
 import PublicLayout from "./layouts/PublicLayout";
 import ProtectedLayout from "./layouts/ProtectedLayout";
-import AdminHomePage from "./pages/AdminHomePage";
+import AdminHomePage from "./pages/admin/AdminHomePage";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import AdminNoticePage from "./pages/AdminNoticePage";
+import AdminNoticePage from "./pages/admin/AdminNoticePage";
 
 const publicRoutes: RouteObject[] = [
   {
