+++ src/pages/myPage/PointPage.tsx
